New IR control file
This file uses functions available from serialRobot and AD.c to get an IR reading and send it via serial to wherever. Future work includes converting ADC values to an IR range in cm. A high degree polynomial needs to be attached to the datapoints.

Currently a main function is called, but the code can be condensed (without a while loop) into another function which can be called from the main control agorithm (supposedly the PID control algorithm).